Create Retest::VersionControl::Git#diff_files
