+++ test/retest/version_control/no_version_control_test.rb
@@ -1,0 +1,12 @@
+require 'test_helper'
+require_relative 'shared_interface'
+
+module Retest
+  class VersionControl::NoVersionControlTest < Minitest::Test
+    def setup
+      @subject = VersionControl::NoVersionControl.new
+    end
+
+    include VersionControl::SharedInterface
+  end
+end
