+++ test/retest/version_control/shared_interface.rb
@@ -1,0 +1,9 @@
+module Retest
+  module VersionControl::SharedInterface
+    def test_interface
+      assert_respond_to @subject.class, :installed?
+      assert_respond_to @subject, :name
+      assert_respond_to @subject, :files
+    end
+  end
+end
